requirements.txt
streamlit
rembg
Pillow
onnxruntime
numpy
librosa
matplotlib

